chore(cli): slim CLI to a headless harness; pnpm play → TUI (ticket 0030)
The TUI is now the play experience, so demote apps/cli from an
interactive readline client to a thin, deterministic, scriptable engine
harness (src/sim.ts): parse --seed/--seats/--hero args, shuffle from a
mulberry32 seed (no Math.random), play one N-seat hand vs the bots —
taking scripted hero actions, defaulting to the cheapest legal continue
when the script runs out — and print a plain greppable transcript (state,
actions, coach verdicts, result), then exit. The readline LineReader and
play.ts are gone.

De-duplicate the helpers both clients share into a new pure package
@holdem/format: the action-input grammar (parseAction/renderLegal) and
the coach value formatters (pct/signedChips/VERDICT_LABEL), with their
tests moved alongside. Both apps now import from it — the TUI's input.ts
and CoachPanel and the CLI's table.ts drop their flagged copies — so the
two clients can't diverge. (Not added to the coverage gate; presentation,
like the apps.) No duplicate helper definitions remain.

pnpm play → TUI; pnpm sim → the headless harness. README updated; the
ROADMAP already described this end state.
