+++ apps/cli/src/table.ts
@@ -1,7 +1,13 @@
 /**
- * Pure helpers for the terminal hand runner (ticket 0004): the trivial opponent, the
- * human-input parser, and the text rendering. Kept free of any I/O so they can be
- * unit-tested — `play.ts` is the thin readline shell that wires them to a real terminal.
+ * The string renderers for the headless CLI harness (ticket 0030): the table view, the result
+ * block, and the coach feedback block it prints to the transcript. Kept free of any I/O so they can
+ * be unit-tested — `sim.ts` is the thin harness that wires them to `process.stdout`.
+ *
+ * The genuinely-shared, framework-agnostic helpers the harness also uses — the action-input grammar
+ * (`parseAction` / `renderLegal`) and the coach value formatters (`pct` / `signedChips` /
+ * `VERDICT_LABEL`) — now live in the pure `@holdem/format` package, shared with the TUI so the two
+ * clients can never diverge on them (ticket 0030 consolidated the former per-app copies). This file
+ * imports them and keeps only the renderers, which are this harness's own presentation.
  */
 
 import {
@@ -10,86 +16,15 @@ import {
   potTotal,
   isComplete,
   type HandState,
-  type Action,
-  type LegalActions,
   type PlayerState,
 } from '@holdem/engine'
 import type { DecisionVerdict, StartingHandVerdict } from '@holdem/coach'
+import { pct, signedChips, VERDICT_LABEL } from '@holdem/format'
 
-/** Result of parsing a line of human input: a legal action, or a message to reprint. */
-export type ParseResult = { ok: true; action: Action } | { ok: false; error: string }
+/** Re-exported for the harness so it has one import surface (`from './table.js'`). */
+export { parseAction, renderLegal } from '@holdem/format'
 
-/** A one-line, human-readable menu of the legal actions, with amounts. */
-export function renderLegal(legal: LegalActions): string {
-  const parts: string[] = []
-  if (legal.fold) parts.push('(f)old')
-  if (legal.check) parts.push('(k)check')
-  if (legal.call) parts.push(`(c)all ${legal.call.amount}`)
-  if (legal.bet) parts.push(`(b)et ${legal.bet.min}-${legal.bet.max}`)
-  if (legal.raise) parts.push(`(r)aise to ${legal.raise.min}-${legal.raise.max}`)
-  if (legal.bet || legal.raise) parts.push('(a)llin')
-  return parts.join('  ')
-}
-
-/**
- * Parse a line of human input against the legal actions. Accepts single-letter or
- * full-word verbs and an optional amount (`b50`, `b 50`, `bet 50`); for bet/raise a
- * missing amount means the minimum, and `a`/`allin` means the maximum.
- */
-export function parseAction(input: string, legal: LegalActions): ParseResult {
-  const m = input
-    .trim()
-    .toLowerCase()
-    .match(/^([a-z]+)\s*(\d+)?$/)
-  if (!m) return { ok: false, error: 'Could not read that — try again.' }
-  const verb = m[1]!
-  const amount = m[2] === undefined ? null : Number(m[2])
-
-  const illegal = (name: string): ParseResult => ({
-    ok: false,
-    error: `${name} is not legal here.`,
-  })
-
-  switch (verb) {
-    case 'f':
-    case 'fold':
-      return legal.fold ? { ok: true, action: { type: 'fold' } } : illegal('Fold')
-    case 'k':
-    case 'check':
-      return legal.check ? { ok: true, action: { type: 'check' } } : illegal('Check')
-    case 'c':
-    case 'call':
-      return legal.call ? { ok: true, action: { type: 'call' } } : illegal('Call')
-    case 'a':
-    case 'allin':
-    case 'shove':
-      if (legal.bet) return { ok: true, action: { type: 'bet', amount: legal.bet.max } }
-      if (legal.raise) return { ok: true, action: { type: 'raise', amount: legal.raise.max } }
-      return illegal('All-in')
-    case 'b':
-    case 'bet':
-      return amountAction('bet', legal.bet, amount)
-    case 'r':
-    case 'raise':
-      return amountAction('raise', legal.raise, amount)
-    default:
-      return { ok: false, error: `Unknown action "${verb}".` }
-  }
-}
-
-function amountAction(
-  type: 'bet' | 'raise',
-  range: { min: number; max: number } | null,
-  amount: number | null,
-): ParseResult {
-  if (!range) return { ok: false, error: `${cap(type)} is not legal here.` }
-  const to = amount ?? range.min // bare verb means the minimum
-  if (to < range.min || to > range.max) {
-    return { ok: false, error: `${cap(type)} must be to ${range.min}-${range.max}.` }
-  }
-  return { ok: true, action: { type, amount: to } }
-}
-
+/** Capitalise the first letter of a word for a section header (`preflop` → `Preflop`). */
 function cap(s: string): string {
   return s[0]!.toUpperCase() + s.slice(1)
 }
@@ -106,9 +41,10 @@ export function renderState(state: HandState, heroSeat: number): string {
   return lines.join('\n')
 }
 
+/** Render one seat's row: who, (revealed) cards, stack, current bet, and any markers. */
 function renderSeat(state: HandState, p: PlayerState, heroSeat: number): string {
-  const name = p.seat === heroSeat ? 'You' : 'Bot'
-  // Hide the opponent's cards until the hand is over.
+  const name = p.seat === heroSeat ? 'You' : `Bot ${p.seat}`
+  // Hide the opponents' cards until the hand is over.
   const reveal = p.seat === heroSeat || isComplete(state)
   const cards = reveal ? p.holeCards.map(formatCard).join(' ') : '?? ??'
   const marks = [
@@ -120,19 +56,19 @@ function renderSeat(state: HandState, p: PlayerState, heroSeat: number): string 
     .filter(Boolean)
     .join(' ')
   const bet = p.committed > 0 ? `  bet ${p.committed}` : ''
-  return `  ${name.padEnd(3)} [${cards}]  stack ${p.stack}${bet}  ${marks}`.trimEnd()
+  return `  ${name.padEnd(5)} [${cards}]  stack ${p.stack}${bet}  ${marks}`.trimEnd()
 }
 
 /** Render the outcome of a completed hand: the showdown (if any) and the payouts. */
 export function renderResult(state: HandState, heroSeat: number): string {
+  const who = (seat: number): string => (seat === heroSeat ? 'You' : `Bot ${seat}`)
   const lines = ['', `── Result ${'─'.repeat(38)}`]
   if (state.endReason === 'showdown') {
     for (const p of state.players) {
       if (p.status === 'folded') continue
-      const who = p.seat === heroSeat ? 'You' : 'Bot'
       const hv = state.showdownHands[p.seat]
       lines.push(
-        `  ${who}: ${p.holeCards.map(formatCard).join(' ')}  — ${hv ? describeHand(hv) : ''}`,
+        `  ${who(p.seat)}: ${p.holeCards.map(formatCard).join(' ')}  — ${hv ? describeHand(hv) : ''}`,
       )
     }
   } else {
@@ -140,32 +76,9 @@ export function renderResult(state: HandState, heroSeat: number): string {
   }
   for (const p of state.players) {
     const won = state.payouts[p.seat] ?? 0
-    if (won > 0) lines.push(`  ${p.seat === heroSeat ? 'You' : 'Bot'} collect ${won}`)
+    if (won > 0) lines.push(`  ${who(p.seat)} collect ${won}`)
   }
   return lines.join('\n')
-}
-
-/** Format a `0..1` equity/pot-odds fraction as a one-decimal percent, e.g. `0.625 → "62.5%"`. */
-function pct(fraction: number): string {
-  return `${(fraction * 100).toFixed(1)}%`
-}
-
-/** Format a chip EV as a signed number, e.g. `4 → "+4"`, `-1.5 → "-1.5"`, `0 → "0"`. */
-function signedChips(ev: number): string {
-  // Round to one decimal *first* so a near-zero EV renders a bare, unsigned `0` rather than
-  // a misleading signed zero (`-0.04 → "0"`, not `"-0"`; also handles JS negative zero).
-  const rounded = Math.round(ev * 10) / 10
-  if (rounded === 0) return '0'
-  // Trim a trailing `.0` so whole-chip EVs read clean (`+4`, not `+4.0`).
-  const magnitude = Math.abs(rounded).toFixed(1).replace(/\.0$/, '')
-  return rounded < 0 ? `-${magnitude}` : `+${magnitude}`
-}
-
-/** The human-readable headline for each {@link DecisionVerdict.verdict} tag. */
-const VERDICT_LABEL: Readonly<Record<DecisionVerdict['verdict'], string>> = {
-  good: 'Good — your action agreed with the math.',
-  leak: 'Leak — the math pointed the other way.',
-  breakEven: 'Break-even — a coin-flip spot; either way is fine.',
 }
 
 /**
@@ -175,10 +88,10 @@ const VERDICT_LABEL: Readonly<Record<DecisionVerdict['verdict'], string>> = {
  * Always shows the postflop-math view of the spot the hero faced: the estimated equity, the
  * pot-odds threshold it is judged against, the chip EV of calling, the EV-correct action
  * (continue vs. fold), and the good/leak/break-even verdict — all taken verbatim from the
- * {@link DecisionVerdict} the coach computed (the CLI does no math of its own). When the
- * decision was preflop, pass the {@link StartingHandVerdict} too and its starting-hand chart
- * tier + rationale lead the block, mirroring how a learner reaches for the chart first
- * preflop and the pot-odds math postflop.
+ * {@link DecisionVerdict} the coach computed (the harness does no math of its own), formatted with
+ * the shared `@holdem/format` helpers. When the decision was preflop, pass the
+ * {@link StartingHandVerdict} too and its starting-hand chart tier + rationale lead the block,
+ * mirroring how a learner reaches for the chart first preflop and the pot-odds math postflop.
  */
 export function renderCoachFeedback(
   verdict: DecisionVerdict,
