+++ apps/tui/package.json
@@ -12,6 +12,7 @@
     "@holdem/bots": "workspace:*",
     "@holdem/coach": "workspace:*",
     "@holdem/engine": "workspace:*",
+    "@holdem/format": "workspace:*",
     "ink": "^7.0.6",
     "react": "^19.2.7"
   },
