+++ apps/tui/src/components/CoachPanel.tsx
@@ -17,49 +17,16 @@
  *   game continues uninterrupted.
  * - `'none'` — no hero decision graded yet: a dim placeholder.
  *
- * Colour is via Ink `color` props; component tests strip ANSI, so they assert on the rendered text
- * (and on the pure formatters below), never on escape bytes.
+ * Colour is via Ink `color` props; component tests strip ANSI, so they assert on the rendered text,
+ * never on escape bytes. The value formatters it renders with (`pct` / `signedChips` /
+ * `VERDICT_LABEL`) live in the shared `@holdem/format` package, so the coach read is phrased
+ * identically here and in the headless CLI harness (ticket 0030 consolidated the former copies).
  */
 
 import { Box, Text } from 'ink'
 import type { DecisionVerdict } from '@holdem/coach'
+import { pct, signedChips, VERDICT_LABEL } from '@holdem/format'
 import type { CoachResult } from '../model.js'
-
-/**
- * Format a `0..1` equity/pot-odds fraction as a one-decimal percent, e.g. `0.625 → "62.5%"`.
- *
- * NOTE (ticket 0030): this is a deliberate, exact copy of `apps/cli/src/table.ts`'s `pct`. The CLI
- * keeps its own copy today; ticket 0030 owns de-duplicating the percent/signed-chip formatting into
- * a shared home. Until then these two copies must stay byte-for-byte identical — keep this cleanly
- * separable so the move is a lift, not a rewrite.
- */
-export function pct(fraction: number): string {
-  return `${(fraction * 100).toFixed(1)}%`
-}
-
-/**
- * Format a chip EV as a signed number, e.g. `4 → "+4"`, `-1.5 → "-1.5"`, `0 → "0"`.
- *
- * NOTE (ticket 0030): an exact copy of `apps/cli/src/table.ts`'s `signedChips`; see {@link pct}.
- * The bare-`0` handling matters: a near-zero EV renders an unsigned `0` (never `-0`), and whole-chip
- * EVs drop the trailing `.0` (`+4`, not `+4.0`).
- */
-export function signedChips(ev: number): string {
-  // Round to one decimal *first* so a near-zero EV renders a bare, unsigned `0` rather than
-  // a misleading signed zero (`-0.04 → "0"`, not `"-0"`; also handles JS negative zero).
-  const rounded = Math.round(ev * 10) / 10
-  if (rounded === 0) return '0'
-  // Trim a trailing `.0` so whole-chip EVs read clean (`+4`, not `+4.0`).
-  const magnitude = Math.abs(rounded).toFixed(1).replace(/\.0$/, '')
-  return rounded < 0 ? `-${magnitude}` : `+${magnitude}`
-}
-
-/** The human-readable headline for each {@link DecisionVerdict.verdict} tag (mirrors the CLI). */
-const VERDICT_LABEL: Readonly<Record<DecisionVerdict['verdict'], string>> = {
-  good: 'Good — your action agreed with the math.',
-  leak: 'Leak — the math pointed the other way.',
-  breakEven: 'Break-even — a coin-flip spot; either way is fine.',
-}
 
 /** The Ink `color` prop for each verdict tag: green good / red leak / yellow break-even. */
 const VERDICT_COLOR: Readonly<Record<DecisionVerdict['verdict'], string>> = {
