+++ package.json
@@ -6,7 +6,8 @@
   "description": "A Texas Hold'em training PWA — play vs bots, drills, and a deterministic odds/strategy coach.",
   "packageManager": "pnpm@11.6.0",
   "scripts": {
-    "play": "pnpm --filter @holdem/cli play",
+    "play": "pnpm --filter @holdem/tui dev",
+    "sim": "pnpm --filter @holdem/cli sim",
     "test": "vitest run",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage",
