+++ packages/format/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@holdem/format",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "description": "Pure-TS presentation helpers shared by the play clients: the action-input grammar and the coach value formatters.",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc -b"
+  },
+  "dependencies": {
+    "@holdem/coach": "workspace:*",
+    "@holdem/engine": "workspace:*"
+  }
+}
